feat: move page to (home)
