+++ src/__tests__/page.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { render, screen } from "@testing-library/react";
-import Home from "@/app/(pages)/page";
+import Home from "@/app/(pages)/(home)/page";
 
 describe("Home", () => {
   it("should render header", () => {
