+++ src/app/(pages)/(home)/favicon.ico
